Aggregate strongly typed Azure DevOps task APIs
Resolve shared builder, test, documentation, and public API conflicts across the exact-prefix task PR set.


Copilot-Session: 19ec0c92-89d9-407a-9902-619406b467e7
